dmalloc: test_all_dmalloc.sh Initial Commit
It works but Dmalloc doesn't work well so my script doesn't work well.
The output isn't very clean and Dmalloc flags all 9 of my source files
with false positives.  I had to circumvent that.  Even worse, I get a
lot more core dumps than Valgrind and Mtrace.  That adds noise and
confusion to the output.

I still have yet to put some Disclaimers and explanations in the shell
file.  As of now, it works just about the best it's going to work.

Changes to be committed:
	new file:   devops/script/test_all_dmalloc.sh

Changes not staged for commit:
	modified:   Makefile
	modified:   devops/script/dependency_checker.sh
	modified:   src/gimme_mem.c
	modified:   src/gimme_mem.h
	modified:   src/mimo_wrappers.h
	modified:   src/mimo_wrappers_dmalloc.c