electric_fence: Updated GDB script
I think I got the GDB script to sit up and beg.
Now, it only logs to a file, binary output is silenced, and I don't see
any output on stdout or stderr.  Everything is in the file.

Now that I've made *that* happen, I can update the efence shell script
to parse the output.

Changes to be committed:
	modified:   devops/script/automate.gdb