bad_code14: WARNING! Do not merge this branch!
At best, one of these memory debuggers (Electric Fence) doesn't find any memory leak.
The rest of the memory debuggers come up with false positives.
I updated the README Results table but not the Conclusion.  We're not
using these results.

BLUF: Don't merge this branch into development.

BACKGROUND: I left this branch as is to document results of running
memory debuggers on a memory leak in a detached thread.

Changes to be committed:
	modified:   README.md