ASAN: Refactored *_code11.c
I was getting some odd results with ASAN and bad_11.bin.
The shell script was reporting everything good.
Manual execution was printing out ASAN "you dun goofed" output.
The stderr output was being silenced in the script and the child
processes exit code was being lost.  I refactored good and bad 11.  The
parent process now waits for the child to exit and uses the child's exit
code as its exit code.  Now, the memory debuggers watching the main
thread will see the exit code of the child (which, as it turns out, is
used by some of these tools to communicate something has gone awry).

When I was done refactoring good/bad 11, I believe Valgrind was able to
pickup the bad 11 error as well.

TO DO: DON'T DO NOW... Propagate this to the other multi-threaded source
files in MIMO!

Changes to be committed:
	modified:   src/bad_code11.c
	modified:   src/good_code11.c